AES test battery quieter

+++ test/Test-AesCrypto.ps1
@@ -5,10 +5,8 @@
     You may copy, modify or distribute this software under the terms of the GNU Public License 2.0.
 #>
 
-Write-Host "This script tests the AesCrypto.cs implementation"
-Write-Host "These tests should pass in both Windows PowerShell and PowerShell 7 on all operating systems"
+# Import the source code
 
-Write-Host "Loading the type from source"
 $code = Get-Content "$PSScriptRoot\..\module\AesCrypto.cs" -Raw
 if( $PSVersionTable.PSVersion.Major -eq 5 ) {
     Add-Type -TypeDefinition $code -ReferencedAssemblies "System.Security"
@@ -16,69 +14,64 @@ if( $PSVersionTable.PSVersion.Major -eq 5 ) {
     Add-Type -TypeDefinition $code -ReferencedAssemblies "System.Security.Cryptography"
 }
 
-Write-Host "Setting up constants for testing"
+# Generate test data
+
 $password = New-PowerPassRandomPassword -Length 32
 $keyFile = "$PSScriptRoot\aes.key"
 $encryptedFile = "$PSScriptRoot\data.aes"
 $data = [System.Guid]::NewGuid().ToString()
 $dataBytes = [System.Text.Encoding]::UTF8.GetBytes($data)
 
-Write-Host "Testing explicit key generation"
+# Test - key generation, writing and reading keys from disk
+
 $aes = New-Object -TypeName "PowerPass.AesCrypto"
 $aes.GenerateKey()
 $secret = [System.Text.Encoding]::UTF8.GetBytes($password)
 $aes.WriteKeyToDisk( $keyFile, [ref] $secret )
 $secret = [System.Text.Encoding]::UTF8.GetBytes($password)
 $aes.ReadKeyFromDisk( $keyFile, [ref] $secret )
+$aes.Dispose()
 
-Write-Host "Clearing key file from disk"
+# Delete previous keys
+
 if( Test-Path $keyFile ) {
     Remove-Item $keyFile -Force
 }
 
-Write-Host "Testing automatic key generation"
-$aes = New-Object -TypeName "PowerPass.AesCrypto"
-$secret = [System.Text.Encoding]::UTF8.GetBytes($password)
-$aes.WriteKeyToDisk( $keyFile, [ref] $secret )
-$secret = [System.Text.Encoding]::UTF8.GetBytes($password)
-$aes.ReadKeyFromDisk( $keyFile, [ref] $secret )
+# Test - encryption and decryption
 
-Write-Host "Testing dispose"
-$aes = New-Object -TypeName "PowerPass.AesCrypto"
-$aes.Dispose()
-$aes = New-Object -TypeName "PowerPass.AesCrypto"
-$aes.GenerateKey()
-$aes.Dispose()
-
-Write-Host "Testing encryption/decryption: " -NoNewline
 $aes = New-Object -TypeName "PowerPass.AesCrypto"
 $aes.GenerateKey()
 $aes.Encrypt( $dataBytes, $encryptedFile )
 $checkDataBytes = $aes.Decrypt( $encryptedFile )
 $checkData = [System.Text.Encoding]::UTF8.GetString($checkDataBytes)
-if( [System.String]::Equals( $data, $checkData, "Ordinal" ) ) {
-    Write-Host "Assert passed"
-} else {
-    Write-Warning "Assert failed"
+if( -not ([System.String]::Equals( $data, $checkData, "Ordinal" )) ) {
+    Write-Warning "Test failed: decryption"
 }
+$aes.Dispose()
 
-Write-Host "Testing decryption with loaded key: " -NoNewline
+# Test - write key then encrypt
+
 $aes = New-Object -TypeName "PowerPass.AesCrypto"
 $secret = [System.Text.Encoding]::UTF8.GetBytes($password)
 $aes.WriteKeyToDisk( $keyFile, [ref] $secret )
 $aes.Encrypt( $dataBytes, $encryptedFile )
+$aes.Dispose()
+
+# Test - read key then decrypt
+
 $aes = New-Object -TypeName "PowerPass.AesCrypto"
 $secret = [System.Text.Encoding]::UTF8.GetBytes($password)
 $aes.ReadKeyFromDisk( $keyFile, [ref] $secret )
 $checkDataBytes = $aes.Decrypt( $encryptedFile )
 $checkData = [System.Text.Encoding]::UTF8.GetString($checkDataBytes)
-if( [System.String]::Equals( $data, $checkData, "Ordinal" ) ) {
-    Write-Host "Assert passed"
-} else {
-    Write-Warning "Assert failed"
+if( -not ([System.String]::Equals( $data, $checkData, "Ordinal" )) ) {
+    Write-Warning "Test failed: read key from disk"
 }
+$aes.Dispose()
 
-Write-Host "Testing large file encryption/decryption: " -NoNewline
+# Test - read large file then encrypt and decrypt
+
 $data = Get-Content "$PSScriptRoot\thelasttrain.txt" -Raw
 $dataBytes = [System.Text.Encoding]::UTF8.GetBytes($data)
 $aes = New-Object -TypeName "PowerPass.AesCrypto"
@@ -87,16 +80,11 @@ $aes.GenerateKey()
 $aes.Encrypt( $dataBytes, $encryptedFile )
 $checkDataBytes = $aes.Decrypt( $encryptedFile )
 $checkData = [System.Text.Encoding]::UTF8.GetString($checkDataBytes)
-if( [System.String]::Equals( $data, $checkData, "Ordinal" ) ) {
-    Write-Host "Assert passed"
-} else {
-    Write-Warning "Assert failed"
+if( -not ([System.String]::Equals( $data, $checkData, "Ordinal" )) ) {
+    Write-Warning "Test failed: large file encrypt and decrypt"
 }
 
-Write-Host "Clean up"
-if( Test-Path $keyFile ) {
-    Remove-Item $keyFile -Force
-}
-if( Test-Path $encryptedFile ) {
-    Remove-Item $encryptedFile -Force
-}
+# Clean up temporary files
+
+if( Test-Path $keyFile ) { Remove-Item $keyFile -Force }
+if( Test-Path $encryptedFile ) { Remove-Item $encryptedFile -Force }
